# Conflicts:
#	src/META-INF/plugin.xml

+++ src/com/dci/intellij/dbn/connection/ConnectionHandlerStatus.java
@@ -10,10 +10,10 @@ public enum ConnectionHandlerStatus implements Property {
     VALID,
     BUSY;
 
-    private final int index = Property.idx(this);
+    private final long index = Property.idx(this);
 
     @Override
-    public int index() {
+    public long index() {
         return index;
     }
 }
